change `[  ]_url` helpers to `[  ]_path` helpers
Closes #45.

The way that we currently have set up the specs and route helper
methods, I think it would be hard to really test for `[  ]_url` helpers
that dynamically change in response to whatever server is hosting the
app. So I kept it simple and just changed the language to refer to path
helpers instead of URL helpers. I will change the wording of the bonus
instructions, as well.